+++ src/com/kaylerrenslow/armaDialogCreator/gui/img/ImagePaths.java
@@ -24,6 +24,5 @@ public class ImagePaths {
 	public static final String ICON_COMPOSITE = prefix + "icons/composite.png";
 	public static final String ICON_UNDO = prefix + "icons/undo.png";
 	public static final String ICON_REDO = prefix + "icons/redo.png";
-	
-	
+	public static final String ICON_HEART = prefix + "icons/heart.png";
 }